requirements.txt
pandas==2.2.3
openpyxl==3.1.5
python-dotenv==1.0.1
pyTelegramBotAPI==4.26.0

Pillow==11.1.0
pyinstaller==6.11.1
pystray=0.19.5
